requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
audioread==3.0.1
av==14.3.0
babel==2.17.0
bcrypt==4.3.0
beam-client==0.2.153
beta9==0.1.191
betterproto-beta9==2.0.1
black==25.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.2
coqpit-config==0.2.0
coqui-tts==0.26.0
coqui-tts-trainer==0.2.3
croniter==3.0.4
cryptography==44.0.3
ctranslate2==4.6.0
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dateparser==1.1.8
decorator==5.2.1
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
fastapi==0.115.12
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
grpcio==1.69.0
grpclib==0.4.8
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
inflect==7.0.0
isort==5.13.2
Jinja2==3.1.6
joblib==1.5.0
jsonlines==1.2.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
monotonic-alignment-search==0.1.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
murmurhash==1.0.12
mypy_extensions==1.1.0
networkx==3.4.2
num2words==0.5.14
numba==0.61.2
numpy==1.26.4
onnxruntime==1.19.2
packaging==25.0
paramiko==3.5.1
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
preshed==3.0.9
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==4.25.7
psutil==7.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyNaCl==1.5.0
pyparsing==3.2.3
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.3.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
SudachiDict-core==20250129
SudachiPy==0.6.10
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
tqdm==4.67.1
transformers==4.51.3
typeguard==2.13.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
